fx(rebase): to be up to date with upstream
bg: won't treat an empty cart as an error & remove the required middleware to add

ft: #184361206 Setup docker for backend application

This changes setup the following services on local development:

1. express server
2. postgress db.

ft(product): add product view by ID [#184361222]

fix(utilsJWT): fix token expiry for password reset [#184361212]

fx:admin get orders #184361232

ft: add payment using Stripe API

fx(rebase): to be up to date with upstream

ft: #184361213 add Two Factor Authentication for Sellers

add all tests for this feature plus it's documentation

GIVEN there is a request to log in a user to the website
WHEN 2FA is enabled on the particular user
THEN Further request for a 2FA token should be requested and once the token is received and validated, the user is allowed to log in.

bg: won't treat an empty cart as an error & remove the required middleware to add

ft: #184361206 Setup docker for backend application

This changes setup the following services on local development:

1. express server
2. postgress db.

ft(product): add product view by ID [#184361222]

fx(rebase): to be up to date with upstream

ft: #184361206 Setup docker for backend application

This changes setup the following services on local development:

1. express server
2. postgress db.

ft(product): add product view by ID [#184361222]

fx(rebase): to be up to date with upstream

bg: won't treat an empty cart as an error & remove the required middleware to add

ft: #184361206 Setup docker for backend application

This changes setup the following services on local development:

1. express server
2. postgress db.

fx(rebase): to be up to date with upstream

Fix(Hound CI) fixing houndCI Error

ft:seller delete specific Item in collection

ft: #184361213 add Two Factor Authentication for Sellers

add all tests for this feature plus it's documentation

GIVEN there is a request to log in a user to the website
WHEN 2FA is enabled on the particular user
THEN Further request for a 2FA token should be requested and once the token is received and validated, the user is allowed to log in.

bg: won't treat an empty cart as an error & remove the required middleware to add

fx(rebase): to be up to date with upstream

ft: #184361213 add Two Factor Authentication for Sellers

add all tests for this feature plus it's documentation

GIVEN there is a request to log in a user to the website
WHEN 2FA is enabled on the particular user
THEN Further request for a 2FA token should be requested and once the token is received and validated, the user is allowed to log in.

bg: won't treat an empty cart as an error & remove the required middleware to add

fx(rebase): to be up to date with upstream