ft: admin set roles and permissions
ft(notification): read all/a notification by ID [#184361235]

fx:change the way to get orders #184361232

ft: admin set roles and permissions

fx(rebase): to be up to date with upstream

Fix(Hound CI) fixing houndCI Error

ft:seller delete specific Item in collection

ft: #184361213 add Two Factor Authentication for Sellers

add all tests for this feature plus it's documentation

GIVEN there is a request to log in a user to the website
WHEN 2FA is enabled on the particular user
THEN Further request for a 2FA token should be requested and once the token is received and validated, the user is allowed to log in.

bg: won't treat an empty cart as an error & remove the required middleware to add

fx(rebase): to be up to date with upstream

ft: #184361213 add Two Factor Authentication for Sellers

add all tests for this feature plus it's documentation

GIVEN there is a request to log in a user to the website
WHEN 2FA is enabled on the particular user
THEN Further request for a 2FA token should be requested and once the token is received and validated, the user is allowed to log in.

bg: won't treat an empty cart as an error & remove the required middleware to add

fx(rebase): to be up to date with upstream

ft: admin set roles and permissions